# setu_openapi.json
{"openapi":"3.1.0","info":{"title":"SetuHQ Solutions Engineer Assignment","version":"0.1.0"},"paths":{"/events":{"post":{"summary":"Ingest Event","operationId":"ingest_event_events_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/transactions":{"get":{"summary":"List Transactions","operationId":"list_transactions_transactions_get","parameters":[{"name":"merchant_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Merchant Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"description":"Filter by created_at >= start_date (ISO 8601)","title":"Start Date"},"description":"Filter by created_at >= start_date (ISO 8601)"},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"description":"Filter by created_at <= end_date (ISO 8601)","title":"End Date"},"description":"Filter by created_at <= end_date (ISO 8601)"},{"name":"sort_by","in":"query","required":false,"schema":{"enum":["created_at","latest_event_timestamp","amount"],"type":"string","description":"Field to sort by","default":"created_at","title":"Sort By"},"description":"Field to sort by"},{"name":"sort_dir","in":"query","required":false,"schema":{"enum":["asc","desc"],"type":"string","description":"Sort direction: asc or desc","default":"desc","title":"Sort Dir"},"description":"Sort direction: asc or desc"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"},"title":"Response List Transactions Transactions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/transactions/{transaction_id}":{"get":{"summary":"Get Transaction","operationId":"get_transaction_transactions__transaction_id__get","parameters":[{"name":"transaction_id","in":"path","required":true,"schema":{"type":"string","title":"Transaction Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reconciliation/summary":{"get":{"summary":"Reconciliation Summary","operationId":"reconciliation_summary_reconciliation_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/reconciliation/discrepancies":{"get":{"summary":"Reconciliation Discrepancies","operationId":"reconciliation_discrepancies_reconciliation_discrepancies_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"EventCreate":{"properties":{"event_id":{"type":"string","title":"Event Id"},"event_type":{"$ref":"#/components/schemas/EventType"},"transaction_id":{"type":"string","title":"Transaction Id"},"merchant_id":{"type":"string","title":"Merchant Id"},"merchant_name":{"type":"string","title":"Merchant Name"},"amount":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Amount"},"currency":{"type":"string","title":"Currency"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["event_id","event_type","transaction_id","merchant_id","merchant_name","amount","currency","timestamp"],"title":"EventCreate"},"EventResponse":{"properties":{"event_id":{"type":"string","title":"Event Id"},"event_type":{"$ref":"#/components/schemas/EventType"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["event_id","event_type","timestamp"],"title":"EventResponse"},"EventType":{"type":"string","enum":["payment_initiated","payment_processed","payment_failed","settled"],"title":"EventType"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MerchantResponse":{"properties":{"merchant_id":{"type":"string","title":"Merchant Id"},"merchant_name":{"type":"string","title":"Merchant Name"}},"type":"object","required":["merchant_id","merchant_name"],"title":"MerchantResponse"},"TransactionResponse":{"properties":{"transaction_id":{"type":"string","title":"Transaction Id"},"merchant_id":{"type":"string","title":"Merchant Id"},"merchant":{"anyOf":[{"$ref":"#/components/schemas/MerchantResponse"},{"type":"null"}]},"amount":{"type":"string","title":"Amount"},"currency":{"type":"string","title":"Currency"},"current_status":{"$ref":"#/components/schemas/EventType"},"latest_event_timestamp":{"type":"string","format":"date-time","title":"Latest Event Timestamp"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"events":{"items":{"$ref":"#/components/schemas/EventResponse"},"type":"array","title":"Events","default":[]}},"type":"object","required":["transaction_id","merchant_id","amount","currency","current_status","latest_event_timestamp"],"title":"TransactionResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}